Fixed a few validation errors that were causing issues in the tests
